+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Each change log are assumed to be accompanied by associated unit test and docume
 
 - added `show()` method to controller in order to support change to the single panel directive
 - changes based on sass framework updates
+- converted to using template instead of templateUrl
+- converted to using child scope instead of isolate scope
 
 ## 0.5.0